Updates based on Pull Request #15 comments - Removed commented lines from build.gradle.kts - Removed extra lines from gradle.properties - Changed Gradle back to 7.2 in gradle-wrapper.properties - Fixed yarn.lock - Reformatted code to match style - Restructured Tests   - Use actual/expect to inject ITestLogcatLogger   - Pull platform-agnostic tests to CommonLogcatTest.kt   - Put platform-specific tests in [Platform]LogcatTest.kt
