+++ gradle.properties
@@ -23,6 +23,3 @@ POM_LICENCE_DIST=repo
 
 POM_DEVELOPER_ID=square
 POM_DEVELOPER_NAME=Square, Inc.
-POM_ARTIFACT_ID=logcat
-POM_NAME=Logcat
-POM_PACKAGING=aar
