+++ src/commonMain/kotlin/logcat/PrintLogger.kt
@@ -3,11 +3,15 @@ package logcat
 /**
  * A [LogcatLogger] that always logs and delegates to [println] concatenating
  * the tag and message, separated by a space. Alternative to [AndroidLogcatLogger]
- * when running on a JVM.
+ * when running on a JVM/JS/Native.
  */
 object PrintLogger : LogcatLogger {
 
-  override fun log(priority: LogPriority, tag: String, message: String) {
+  override fun log(
+    priority: LogPriority,
+    tag: String,
+    message: String
+  ) {
     println("$tag $message")
   }
 }
