+++ src/commonMain/kotlin/logcat/defaultThrowableAsLog.kt
@@ -1,6 +1,6 @@
 package logcat
 
 fun Throwable.defaultThrowableAsLog(): String {
-    printStackTrace()
-    return stackTraceToString()
+  printStackTrace()
+  return stackTraceToString()
 }
