+++ src/commonTest/kotlin/logcat/ITestLogcatLogger.kt
@@ -1,0 +1,9 @@
+package logcat
+
+/**
+ * Extension of the LogcatLogger interface provided in the logcat library that
+ * provides access to `latestLog` for inspection during tests.
+ */
+interface ITestLogcatLogger: LogcatLogger {
+  var latestLog: Log?
+}
