+++ src/commonTest/kotlin/logcat/Log.kt
@@ -1,0 +1,7 @@
+package logcat
+
+data class Log(
+  val priority: LogPriority,
+  val tag: String,
+  val message: String
+)
