+++ src/jsTest/kotlin/logcat/platformTestLogger.kt
@@ -1,0 +1,3 @@
+package logcat
+
+actual fun platformTestLogger(isLoggable: (LogPriority) -> Boolean): ITestLogcatLogger = TestLogcatLogger(isLoggable = isLoggable)
