+++ src/nativeMain/kotlin/logcat/outerClassSimpleNameInternalOnlyDoNotUseKThxBye.kt
@@ -1,20 +1,15 @@
 package logcat
 
 private fun String?.stripCompanions(): String? {
-    return if (this.isNullOrEmpty() || this == "Companion") null else this
+  return if (this.isNullOrEmpty() || this == "Companion") null else this
 }
 
 actual fun Any.outerClassSimpleNameInternalOnlyDoNotUseKThxBye(): String {
-    val nativeClass = this::class
-    val fullClassName = nativeClass.simpleName.stripCompanions() ?: "$nativeClass"
-    val outerClassName = fullClassName.removePrefix("class ").substringBefore('$')
-    val simplerOuterClassName = outerClassName.removeSuffix(".Companion").substringAfterLast('.')
-
-    println("nativeClass = $nativeClass")
-    println("fullClassName = $fullClassName")
-    println("outerClassName = $outerClassName")
-    println("simpleOuterClassName = $simplerOuterClassName")
-    return simplerOuterClassName.stripCompanions()
-        ?: fullClassName.stripCompanions()
-        ?: "Companion"
+  val nativeClass = this::class
+  val fullClassName = nativeClass.simpleName.stripCompanions() ?: "$nativeClass"
+  val outerClassName = fullClassName.removePrefix("class ").substringBefore('$')
+  val simplerOuterClassName = outerClassName.removeSuffix(".Companion").substringAfterLast('.')
+  return simplerOuterClassName.stripCompanions()
+    ?: fullClassName.stripCompanions()
+    ?: "Companion"
 }
